pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = "IITM"
authors = ["Govind Soni <gsoni.2301@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.0"
reportlab = "^4.2.5"


[tool.poetry.group.dev.dependencies]
flower = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
